Update CoinFlip.sol
Added fail event, and requiring amounts to be sent to contract. This is to make sure the owner and user are both betting equal and fair amounts. There is still room for improvement with RNG as well as letting the user bet more money. Note that uint bet will be in weis not ether.